refactor: fold now-HMM-only release helpers into hmm package
Removing the remote-reference feature left virtool/github.py and
virtool/releases.py with HMM as their only consumer. Move
create_update_subdocument into virtool/hmm/utils.py and the release
manifest fetcher into virtool/hmm/releases.py, then delete the two
top-level modules. Drop the now-dead ReleaseType.REFERENCES member and
the unused ReleaseManifest dataclasses.
